Cleaned up the folders
