--- src/controller/tcp/TCPListener.java
@@ -1,5 +1,0 @@
-package controller.tcp;
-
-public interface TCPListener {
-	
-}
