--- src/networking/udp/UDPSocket.java
@@ -1,14 +1,0 @@
-package networking.udp;
-
-import java.net.DatagramPacket;
-
-public class UDPSocket {
-	
-	
-	
-	public void onPacketReceived(DatagramPacket inFromClient) {
-		
-	}
-	
-	
-}
