--- src/old/NetworkListener.java
@@ -1,7 +1,0 @@
-
-public interface NetworkListener {
-	
-	public boolean tcpOpen();
-	public void startTCP();
-	
-}
